+++ src/App.tsx
@@ -1,17 +1,17 @@
 import './App.css';
 import { RecoilRoot } from 'recoil';
 import { Routes, Route, BrowserRouter } from 'react-router-dom';
-import DetailPetPage from 'pages/detailPet/DetailPetPage';
 import Home from 'pages/Home';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import ProfileListPage from 'pages/ProfileListPage';
-import ShelterInfoPage from 'pages/ShelterInfoPage';
-import UrgentListPage from 'pages/UrgentListPage';
-import NewListPage from 'pages/NewListPage';
+import DetailPetPage from 'pages/detailPet/DetailPetPage';
+import ProfileListPage from 'pages/ProfileList/ProfileListPage';
 import LoginPage from 'pages/LoginPage';
+import MapPage from 'pages/map/MapPage';
+import NewListPage from 'pages/ProfileList/newList/NewListPage';
+import RegisterPage from 'pages/RegisterPage';
+import ShelterInfoPage from 'pages/ShelterInfoPage';
 import SignupPage from 'pages/SignupPage';
-import RegisterPage from 'pages/register/RegisterPage';
-import MapPage from 'pages/MapPage';
+import UrgentListPage from 'pages/ProfileList/urgentList/UrgentListPage';
 
 const queryClient = new QueryClient();
