+++ src/pages/ProfileList/ProfileListPage.tsx
@@ -1,4 +1,4 @@
-import ProfileListHome from 'components/organisms/ProfileListHome';
+import ProfileListHome from 'pages/ProfileList/ProfileListHome';
 import GNB from 'layouts/GNB';
 
 const ProfileListPage = () => {
