+++ src/pages/RegisterPage.tsx
@@ -1,5 +1,5 @@
 import DayModalInput from 'pages/register/DayModalInput';
-import MRegisterForm from 'components/organisms/MRegisterForm';
+import MRegisterForm from 'pages/register/MRegisterForm';
 import GNB from 'layouts/GNB';
 import StatusSelectGroup from './register/StatusSelectGroup';
 import RegisterHeader from './register/RegisterHeader';
