+++ src/pages/register/DetailRadio.tsx
