+++ src/pages/register/MRegisterForm.tsx
@@ -1,5 +1,4 @@
 import { useRecoilState } from 'recoil';
-import { useState } from 'react';
 import registerState from 'recoil/registerState';
 import VMRegisterForm from './VMRegisterForm';
