+++ src/pages/register/RegisterHeader.tsx
@@ -2,7 +2,9 @@ import { useState, useRef } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import ImageVideoInput from 'pages/register/ImageVideoInput';
 import ModalPortal from 'commons/modals/ModalPortal';
-import RegisterModal, { RegisterModalProps } from 'commons/modals/RegisterModal';
+import RegisterModal, {
+  RegisterModalProps,
+} from 'commons/modals/RegisterModal';
 import { useRecoilValue } from 'recoil';
 import registerState from 'recoil/registerState';
 import { useNavigate } from 'react-router-dom';
