--- src/pages/register/RegisterPage.tsx
@@ -1,19 +1,0 @@
-import Radio from 'commons/Radio';
-import DayModalInput from 'pages/register/DayModalInput';
-import MRegisterForm from 'components/organisms/MRegisterForm';
-import RegisterHeader from 'pages/register/RegisterHeader';
-import GNB from 'layouts/GNB';
-
-const RegisterPage = () => {
-  return (
-    <div>
-      <GNB />
-      <RegisterHeader />
-      <MRegisterForm />
-      <DayModalInput />
-      <Radio />
-    </div>
-  );
-};
-
-export default RegisterPage;
